move some code to controllers
