+++ platform-operator/controllers/component/component.go
