+++ platform-operator/controllers/component/helm_component_test.go
