+++ platform-operator/controllers/component/istio_test.go
