+++ platform-operator/controllers/component/registry.go
