+++ platform-operator/controllers/component/registry_test.go
