+++ platform-operator/controllers/component/testdata/Chart.yaml
