+++ platform-operator/controllers/component/verrazzano_test.go
