+++ platform-operator/controllers/installjob/configmap.go
