+++ platform-operator/controllers/installjob/configmap_test.go
