+++ platform-operator/controllers/installjob/install_config.go
