+++ platform-operator/controllers/installjob/job_test.go
