+++ platform-operator/controllers/uninstalljob/job.go
