+++ platform-operator/controllers/uninstalljob/job_test.go
